feat: add changes in  docker and scripts in mysql-init (#21)
