+++ mysql-init/00-create-database.sql
@@ -1,0 +1,2 @@
+CREATE DATABASE IF NOT EXISTS agenda;
+USE agenda;
