--- src/main/java/infrastructure/sql/connection/DBconection.java
@@ -1,21 +1,0 @@
-import java.io.InputStream;
-import java.sql.Connection;
-import java.sql.DriverManager;
-import java.util.Properties;
-
-public class DBConnection {
-
-    public static Connection getConnection() throws Exception {
-        Properties props = new Properties();
-
-        try (InputStream input = DBConnection.class.getClassLoader().getResourceAsStream("db.properties")) {
-            props.load(input);
-        }
-
-        String url = props.getProperty("db.url");
-        String user = props.getProperty("db.user");
-        String password = props.getProperty("db.password");
-
-        return DriverManager.getConnection(url, user, password);
-    }
-}
