+++ src/main/resources/db.properties
@@ -1,3 +1,3 @@
-db.url=jdbc:mysql://localhost:3307/my_agend
+db.url=jdbc:mysql://localhost:3307/agenda?useSSL=false&serverTimezone=UTC
 db.user=root
 db.password=pajoan
